fix some bugs

+++ tf_model_zoo/ECO/layer_factory.py
@@ -15,7 +15,7 @@ def get_basic_layer(info, channels=None, conv_bias=False, num_segments=4):
 
     attr = info['attrs'] if 'attrs' in info else list()
     if id=="res5b_pool":
-        attr['kernel_d'] = num_segments / 4
+        attr['kernel_d'] = int(num_segments / 4)
 
     out, op, in_vars = parse_expr(info['expr'])
     assert(len(out) == 1)
